backend/requirements.txt
absl-py==0.10.0
asgiref==3.2.10
astunparse==1.6.3
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
channels==3.0.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.0
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
eunjeon==0.4.0
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.33.1
h5py==2.10.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
joblib==0.17.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.3
mysqlclient==2.0.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.4
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.51.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.11
vue==0.0.1
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.2.0
celery==4.3.0
vine==1.3.0
